chatbots/website_chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
sentence-transformers
groq
pydantic
python-multipart
starlette
pymupdf
langchain-text-splitters
torch>=2.4.0
